Time: 86 ms (91.25%), Space: 91.3 MB (96.02%) - LeetHub

+++ 3066-minimum-operations-to-exceed-threshold-value-ii/3066-minimum-operations-to-exceed-threshold-value-ii.cpp
@@ -5,8 +5,8 @@ public:
 
         int op = 0;
         while(pq.size() >= 2 && pq.top() < k) {
-            int x = pq.top(); pq.pop();
-            int y = pq.top(); pq.pop();
+            long long x = pq.top(); pq.pop();
+            long long y = pq.top(); pq.pop();
             pq.push(min(x, y) * 2LL + max(x, y));
             op++;
         }
